fix(ui): polish navigation history popout
