+++ src/renderer/components/TopNav/TopNav.scss
@@ -83,6 +83,12 @@
   user-select: none;
 }
 
+.navIconButton :deep(.iconDropdown) {
+  box-shadow:
+    0 16px 40px rgb(0 0 0 / 55%),
+    0 3px 10px rgb(0 0 0 / 45%);
+}
+
 .navButton {
   border-radius: 50%;
   border-style: none;
@@ -282,4 +288,3 @@
     inset-inline-start: var(--vertical-tab-bar-width, 220px);
   }
 }
-
